+++ src/Typin/Typin.Core/Internal/Extensions/TypeExtensions.cs
@@ -13,13 +13,6 @@
             return type.GetInterfaces().Contains(interfaceType);
         }
 
-        public static Type? TryGetEnumerableArgumentUnderlyingType(this PropertyInfo? property)
-        {
-            return property is not null && property.PropertyType != typeof(string)
-                       ? property.PropertyType.TryGetEnumerableUnderlyingType()
-                       : null;
-        }
-
         public static Type? TryGetNullableUnderlyingType(this Type type)
         {
             return Nullable.GetUnderlyingType(type);
@@ -41,6 +34,40 @@
                        .Where(t => t is not null)
                        .OrderByDescending(t => t != typeof(object)) // prioritize more specific types
                        .FirstOrDefault();
+        }
+
+        public static MethodInfo GetToStringMethod(this Type type)
+        {
+            // ToString() with no params always exists
+            return type.GetMethod(nameof(ToString), Type.EmptyTypes)!;
+        }
+
+        public static bool IsToStringOverriden(this Type type)
+        {
+            return type.GetToStringMethod() != typeof(object).GetToStringMethod();
+        }
+
+        public static MethodInfo? TryGetStaticParseMethod(this Type type, bool withFormatProvider = false)
+        {
+            Type[] argumentTypes = withFormatProvider
+                ? new[] { typeof(string), typeof(IFormatProvider) }
+                : new[] { typeof(string) };
+
+            return type.GetMethod("Parse",
+                                  BindingFlags.Public | BindingFlags.Static,
+                                  null,
+                                  argumentTypes,
+                                  null);
+        }
+
+        public static Array ToNonGenericArray<T>(this IEnumerable<T> source, Type elementType)
+        {
+            ICollection sourceAsCollection = source as ICollection ?? source.ToArray();
+
+            Array array = Array.CreateInstance(elementType, sourceAsCollection.Count);
+            sourceAsCollection.CopyTo(array, 0);
+
+            return array;
         }
     }
 }