+++ src/Typin/Typin.Tests/Data/Commands/Invalid/InvalidParameterConverterCommand.cs
@@ -1,0 +1,16 @@
+﻿namespace Typin.Tests.Data.Commands.Valid
+{
+    using Typin.Attributes;
+    using Typin.Tests.Data.Commands;
+    using Typin.Tests.Data.CustomTypes.InitializableByConverter;
+
+    [Command("cmd")]
+    public class InvalidParameterConverterCommand : SelfSerializeCommandBase
+    {
+        [CommandParameter(0, Converter = typeof(object))]
+        public InitializableClassTypeByConverter? StringInitializable { get; init; }
+
+        [CommandOption("str-nullable-struct", Converter = typeof(InitializableNullableStructTypeByConverter_Converter))]
+        public InitializableStructTypeByConverter? StringNullableInitializableStruct { get; init; }
+    }
+}