+++ src/Typin/Typin/Internal/Schemas/CommandSchemaResolver.cs
@@ -101,7 +101,7 @@
             }
 
             CommandParameterSchema[]? nonScalarParameters = command.Parameters
-                                                                   .Where(p => !p.IsScalar)
+                                                                   .Where(p => !p.BindableProperty.IsScalar)
                                                                    .ToArray();
 
             if (nonScalarParameters.Length > 1)
@@ -115,7 +115,7 @@
             CommandParameterSchema? nonLastNonScalarParameter = command.Parameters
                                                                        .OrderByDescending(a => a.Order)
                                                                        .Skip(1)
-                                                                       .LastOrDefault(p => !p.IsScalar);
+                                                                       .LastOrDefault(p => !p.BindableProperty.IsScalar);
 
             if (nonLastNonScalarParameter is not null)
             {